+++ src/components/level/index.ts
@@ -1,5 +1,5 @@
 export * from "./agent";
-export * from "./navigation";
+export * from "./pathfinder";
 export * from "./patrol";
 export * from "./sentry";
 export * from "./tile";